Moved, mnet, mpacket and nx into common sub folder
