+++ channel/item.go
@@ -10,9 +10,9 @@ import (
 	"time"
 
 	"github.com/Hucaru/Valhalla/common"
+	"github.com/Hucaru/Valhalla/common/mpacket"
+	"github.com/Hucaru/Valhalla/common/nx"
 	"github.com/Hucaru/Valhalla/constant"
-	"github.com/Hucaru/Valhalla/mpacket"
-	"github.com/Hucaru/Valhalla/nx"
 	"github.com/google/uuid"
 )
