+++ channel/message.go
@@ -1,8 +1,8 @@
 package channel
 
 import (
+	"github.com/Hucaru/Valhalla/common/mpacket"
 	"github.com/Hucaru/Valhalla/common/opcode"
-	"github.com/Hucaru/Valhalla/mpacket"
 )
 
 /*
