+++ channel/pools.go
@@ -7,9 +7,9 @@ import (
 	"math/rand"
 	"time"
 
+	"github.com/Hucaru/Valhalla/common/mpacket"
+	"github.com/Hucaru/Valhalla/common/nx"
 	"github.com/Hucaru/Valhalla/common/opcode"
-	"github.com/Hucaru/Valhalla/mpacket"
-	"github.com/Hucaru/Valhalla/nx"
 )
 
 type lifePoolRectangle struct {
