+++ channel/room.go
@@ -6,8 +6,8 @@ import (
 	"math/rand"
 	"time"
 
+	"github.com/Hucaru/Valhalla/common/mpacket"
 	"github.com/Hucaru/Valhalla/common/opcode"
-	"github.com/Hucaru/Valhalla/mpacket"
 )
 
 const (
