+++ channel/script.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Hucaru/Valhalla/mnet"
+	"github.com/Hucaru/Valhalla/common/mnet"
 	"github.com/dop251/goja"
 	"github.com/fsnotify/fsnotify"
 )
