+++ common/mnet/conn.go
@@ -5,9 +5,9 @@ import (
 	"net"
 	"time"
 
-	"github.com/Hucaru/Valhalla/mnet/crypt"
+	"github.com/Hucaru/Valhalla/common/mnet/crypt"
 
-	"github.com/Hucaru/Valhalla/mpacket"
+	"github.com/Hucaru/Valhalla/common/mpacket"
 )
 
 type MConn interface {
