+++ common/mpacket/reader.go
