+++ common/nx/mob.go
