+++ common/nx/skill.go
