+++ login/handlers.go
@@ -7,11 +7,11 @@ import (
 	"strings"
 
 	"github.com/Hucaru/Valhalla/common"
+	"github.com/Hucaru/Valhalla/common/mnet"
+	"github.com/Hucaru/Valhalla/common/mpacket"
 	"github.com/Hucaru/Valhalla/common/opcode"
 	"github.com/Hucaru/Valhalla/constant"
 	"github.com/Hucaru/Valhalla/internal"
-	"github.com/Hucaru/Valhalla/mnet"
-	"github.com/Hucaru/Valhalla/mpacket"
 )
 
 // HandleClientPacket data
