+++ server_login.go
@@ -8,12 +8,12 @@ import (
 	"sync"
 	"time"
 
+	"github.com/Hucaru/Valhalla/common/mpacket"
+	"github.com/Hucaru/Valhalla/common/nx"
 	"github.com/Hucaru/Valhalla/constant"
 	"github.com/Hucaru/Valhalla/login"
-	"github.com/Hucaru/Valhalla/mpacket"
-	"github.com/Hucaru/Valhalla/nx"
 
-	"github.com/Hucaru/Valhalla/mnet"
+	"github.com/Hucaru/Valhalla/common/mnet"
 )
 
 type loginServer struct {
